feat: create context menu items by function and remove ContextMenu component
