+++ src/lib/components/modes/tablemode/contextmenu/createTableContextMenuItems.ts
@@ -1,0 +1,240 @@
+import type { ContextMenuItem, DocumentState, JSONParser } from 'svelte-jsoneditor'
+import {
+  faCheckSquare,
+  faClone,
+  faCopy,
+  faCut,
+  faPaste,
+  faPen,
+  faPlus,
+  faSquare,
+  faTrashCan
+} from '@fortawesome/free-solid-svg-icons'
+import { isKeySelection, isMultiSelection, isValueSelection } from 'svelte-jsoneditor'
+import { compileJSONPointer, getIn } from 'immutable-json-patch'
+import { getFocusPath, singleItemSelected } from '$lib/logic/selection'
+import { isObjectOrArray } from '$lib/utils/typeUtils'
+import { getEnforceString } from '$lib/logic/documentState'
+
+export default function ({
+  json,
+  documentState,
+  parser,
+  onEditValue,
+  onEditRow,
+  onToggleEnforceString,
+  onCut,
+  onCopy,
+  onPaste,
+  onRemove,
+  onDuplicateRow,
+  onInsertBeforeRow,
+  onInsertAfterRow,
+  onRemoveRow
+}: {
+  json: unknown | undefined
+  documentState: DocumentState
+  parser: JSONParser
+  onEditValue: () => void
+  onEditRow: () => void
+  onToggleEnforceString: () => void
+  onCut: (indent: boolean) => void
+  onCopy: (indent: boolean) => void
+  onPaste: () => void
+  onRemove: () => void
+  onDuplicateRow: () => void
+  onInsertBeforeRow: () => void
+  onInsertAfterRow: () => void
+  onRemoveRow: () => void
+}): ContextMenuItem[] {
+  const selection = documentState.selection
+
+  const hasJson = json !== undefined
+  const hasSelection = !!selection
+  const focusValue =
+    json !== undefined && selection ? getIn(json, getFocusPath(selection)) : undefined
+
+  const hasSelectionContents =
+    hasJson &&
+    (isMultiSelection(selection) || isKeySelection(selection) || isValueSelection(selection))
+
+  const canEditValue = hasJson && selection != null && singleItemSelected(selection)
+  const canEnforceString = canEditValue && !isObjectOrArray(focusValue)
+
+  const enforceString =
+    selection != null && focusValue !== undefined
+      ? getEnforceString(
+          focusValue,
+          documentState.enforceStringMap,
+          compileJSONPointer(getFocusPath(selection)),
+          parser
+        )
+      : false
+
+  return [
+    { type: 'separator' },
+    {
+      type: 'row',
+      items: [
+        {
+          type: 'column',
+          items: [
+            { type: 'label', text: 'Table cell:' },
+            {
+              type: 'dropdown-button',
+              main: {
+                type: 'button',
+                onClick: () => onEditValue(),
+                icon: faPen,
+                text: 'Edit',
+                title: 'Edit the value (Double-click on the value)',
+                disabled: !canEditValue
+              },
+              width: '11em',
+              items: [
+                {
+                  type: 'button',
+                  icon: faPen,
+                  text: 'Edit',
+                  title: 'Edit the value (Double-click on the value)',
+                  onClick: () => onEditValue(),
+                  disabled: !canEditValue
+                },
+                {
+                  type: 'button',
+                  icon: enforceString ? faCheckSquare : faSquare,
+                  text: 'Enforce string',
+                  title: 'Enforce keeping the value as string when it contains a numeric value',
+                  onClick: () => onToggleEnforceString(),
+                  disabled: !canEnforceString
+                }
+              ]
+            },
+            {
+              type: 'dropdown-button',
+              main: {
+                type: 'button',
+                onClick: () => onCut(true),
+                icon: faCut,
+                text: 'Cut',
+                title: 'Cut selected contents, formatted with indentation (Ctrl+X)',
+                disabled: !hasSelectionContents
+              },
+              width: '10em',
+              items: [
+                {
+                  type: 'button',
+                  icon: faCut,
+                  text: 'Cut formatted',
+                  title: 'Cut selected contents, formatted with indentation (Ctrl+X)',
+                  onClick: () => onCut(true),
+                  disabled: !hasSelectionContents
+                },
+                {
+                  type: 'button',
+                  icon: faCut,
+                  text: 'Cut compacted',
+                  title: 'Cut selected contents, without indentation (Ctrl+Shift+X)',
+                  onClick: () => onCut(false),
+                  disabled: !hasSelectionContents
+                }
+              ]
+            },
+            {
+              type: 'dropdown-button',
+              main: {
+                type: 'button',
+                onClick: () => onCopy(true),
+                icon: faCopy,
+                text: 'Copy',
+                title: 'Copy selected contents, formatted with indentation (Ctrl+C)',
+                disabled: !hasSelectionContents
+              },
+              width: '12em',
+              items: [
+                {
+                  type: 'button',
+                  icon: faCopy,
+                  text: 'Copy formatted',
+                  title: 'Copy selected contents, formatted with indentation (Ctrl+C)',
+                  onClick: () => onCopy(false),
+                  disabled: !hasSelectionContents
+                },
+                {
+                  type: 'button',
+                  icon: faCopy,
+                  text: 'Copy compacted',
+                  title: 'Copy selected contents, without indentation (Ctrl+Shift+C)',
+                  onClick: () => onCopy(false),
+                  disabled: !hasSelectionContents
+                }
+              ]
+            },
+            {
+              type: 'button',
+              onClick: () => onPaste(),
+              icon: faPaste,
+              text: 'Paste',
+              title: 'Paste clipboard contents (Ctrl+V)',
+              disabled: !hasSelection
+            },
+            {
+              type: 'button',
+              onClick: () => onRemove(),
+              icon: faTrashCan,
+              text: 'Remove',
+              title: 'Remove selected contents (Delete)',
+              disabled: !hasSelectionContents
+            }
+          ]
+        },
+        {
+          type: 'column',
+          items: [
+            { type: 'label', text: 'Table row:' },
+            {
+              type: 'button',
+              onClick: () => onEditRow(),
+              icon: faPen,
+              text: 'Edit row',
+              title: 'Edit the current row',
+              disabled: !hasSelectionContents
+            },
+            {
+              type: 'button',
+              onClick: () => onDuplicateRow(),
+              icon: faClone,
+              text: 'Duplicate row',
+              title: 'Duplicate the current row',
+              disabled: !hasSelection
+            },
+            {
+              type: 'button',
+              onClick: () => onInsertBeforeRow(),
+              icon: faPlus,
+              text: 'Insert before',
+              title: 'Insert a row before the current row',
+              disabled: !hasSelection
+            },
+            {
+              type: 'button',
+              onClick: () => onInsertAfterRow(),
+              icon: faPlus,
+              text: 'Insert after',
+              title: 'Insert a row after the current row',
+              disabled: !hasSelection
+            },
+            {
+              type: 'button',
+              onClick: () => onRemoveRow(),
+              icon: faTrashCan,
+              text: 'Remove row',
+              title: 'Remove current row',
+              disabled: !hasSelection
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
