+++ src/lib/components/modes/treemode/TreeMode.svelte
@@ -2,26 +2,7 @@
 
 <script lang="ts">
   import { createAutoScrollHandler } from '../../controls/createAutoScrollHandler.js'
-  import {
-    faArrowRightArrowLeft,
-    faCaretSquareDown,
-    faCaretSquareUp,
-    faCheck,
-    faClone,
-    faCode,
-    faCopy,
-    faCropAlt,
-    faCut,
-    faFilter,
-    faPaste,
-    faPen,
-    faPlus,
-    faSortAmountDownAlt,
-    faTrashCan,
-    faWrench,
-    faCheckSquare,
-    faSquare
-  } from '@fortawesome/free-solid-svg-icons'
+  import { faCheck, faCode, faWrench } from '@fortawesome/free-solid-svg-icons'
   import { createDebug } from '$lib/utils/debug.js'
   import type { JSONPatchDocument, JSONPath } from 'immutable-json-patch'
   import { compileJSONPointer, existsIn, getIn, immutableJSONPatch } from 'immutable-json-patch'
@@ -91,7 +72,6 @@
     isValueSelection,
     removeEditModeFromSelection,
     selectAll,
-    singleItemSelected,
     hasSelectionContents,
     updateSelectionInDocumentState
   } from '$lib/logic/selection.js'
@@ -116,13 +96,12 @@
     repairPartialJson
   } from '$lib/utils/jsonUtils.js'
   import { keyComboFromEvent } from '$lib/utils/keyBindings.js'
-  import { isObject, isObjectOrArray, isUrl, stringConvert } from '$lib/utils/typeUtils.js'
+  import { isObjectOrArray, isUrl, stringConvert } from '$lib/utils/typeUtils.js'
   import { createFocusTracker } from '../../controls/createFocusTracker.js'
   import Message from '../../controls/Message.svelte'
   import ValidationErrorsOverview from '../../controls/ValidationErrorsOverview.svelte'
   import CopyPasteModal from '../../modals/CopyPasteModal.svelte'
   import JSONRepairModal from '../../modals/JSONRepairModal.svelte'
-  import TreeContextMenu from './contextmenu/TreeContextMenu.svelte'
   import JSONNode from './JSONNode.svelte'
   import TreeMenu from './menu/TreeMenu.svelte'
   import Welcome from './Welcome.svelte'
@@ -182,6 +161,8 @@
   import JSONPreview from '../../controls/JSONPreview.svelte'
   import type { Context } from 'svelte-simple-modal'
   import type { ContextMenuItem } from '$lib/types'
+  import ContextMenu from '../../controls/contextmenu/ContextMenu.svelte'
+  import createTreeContextMenuItems from './contextmenu/createTreeContextMenuItems'
 
   const debug = createDebug('jsoneditor:TreeMode')
 
@@ -231,67 +212,6 @@
   // modalOpen is true when one of the modals is open.
   // This is used to track whether the editor still has focus
   let modalOpen = false
-
-  $: selection = documentState.selection
-
-  $: hasJson = json !== undefined
-  $: hasSelection = !!selection
-  $: rootSelected = selection ? isEmpty(getFocusPath(selection)) : false
-  $: focusValue = selection ? getIn(json, getFocusPath(selection)) : undefined
-  $: editValueText = Array.isArray(focusValue)
-    ? 'Edit array'
-    : isObject(focusValue)
-      ? 'Edit object'
-      : 'Edit value'
-
-  $: canDuplicate = hasJson && hasSelectionContents(documentState.selection) && !rootSelected // must not be root
-
-  $: canExtract =
-    hasJson &&
-    selection != null &&
-    (isMultiSelection(selection) || isValueSelection(selection)) &&
-    !rootSelected // must not be root
-
-  $: canEditKey =
-    hasJson &&
-    selection != null &&
-    singleItemSelected(selection) &&
-    !rootSelected &&
-    !Array.isArray(getIn(json, initial(getFocusPath(selection))))
-
-  $: canEditValue = hasJson && selection != null && singleItemSelected(selection)
-  $: canEnforceString = canEditValue && !isObjectOrArray(focusValue)
-
-  $: convertMode = hasSelectionContents(documentState.selection)
-  $: insertOrConvertText = convertMode ? 'Convert to:' : 'Insert:'
-  $: canInsertOrConvertStructure = convertMode ? false : hasSelection
-  $: canInsertOrConvertObject = convertMode
-    ? canConvert(selection) && !isObject(focusValue)
-    : hasSelection
-  $: canInsertOrConvertArray = convertMode
-    ? canConvert(selection) && !Array.isArray(focusValue)
-    : hasSelection
-  $: canInsertOrConvertValue = convertMode
-    ? canConvert(selection) && isObjectOrArray(focusValue)
-    : hasSelection
-
-  $: enforceString =
-    selection != null && focusValue
-      ? getEnforceString(
-          focusValue,
-          documentState.enforceStringMap,
-          compileJSONPointer(getFocusPath(selection)),
-          parser
-        )
-      : false
-
-  function handleInsertOrConvert(type: InsertType) {
-    if (convertMode) {
-      handleConvert(type)
-    } else {
-      handleInsertFromContextMenu(type)
-    }
-  }
 
   createFocusTracker({
     onMount,
@@ -1887,245 +1807,40 @@
     offsetLeft,
     showTip
   }: AbsolutePopupOptions) {
-    const defaultItems: ContextMenuItem[] = [
-      {
-        type: 'row',
-        items: [
-          {
-            type: 'button',
-            onClick: () => handleEditKey(),
-            icon: faPen,
-            text: 'Edit key',
-            title: 'Edit the key (Double-click on the key)',
-            disabled: !canEditKey
-          },
-          {
-            type: 'dropdown-button',
-            main: {
-              type: 'button',
-              onClick: () => handleEditValue(),
-              icon: faPen,
-              text: editValueText,
-              title: 'Edit the value (Double-click on the value)',
-              disabled: !canEditValue
-            },
-            width: '11em',
-            items: [
-              {
-                type: 'button',
-                icon: faPen,
-                text: editValueText,
-                title: 'Edit the value (Double-click on the value)',
-                onClick: () => handleEditValue(),
-                disabled: !canEditValue
-              },
-              {
-                type: 'button',
-                icon: enforceString ? faCheckSquare : faSquare,
-                text: 'Enforce string',
-                title: 'Enforce keeping the value as string when it contains a numeric value',
-                onClick: () => handleToggleEnforceString(),
-                disabled: !canEnforceString
-              }
-            ]
-          }
-        ]
-      },
-      { type: 'separator' },
-      {
-        type: 'row',
-        items: [
-          {
-            type: 'dropdown-button',
-            main: {
-              type: 'button',
-              onClick: () => handleCut(true),
-              icon: faCut,
-              text: 'Cut',
-              title: 'Cut selected contents, formatted with indentation (Ctrl+X)',
-              disabled: !hasSelectionContents
-            },
-            width: '10em',
-            items: [
-              {
-                type: 'button',
-                icon: faCut,
-                text: 'Cut formatted',
-                title: 'Cut selected contents, formatted with indentation (Ctrl+X)',
-                onClick: () => handleCut(true),
-                disabled: !hasSelectionContents
-              },
-              {
-                type: 'button',
-                icon: faCut,
-                text: 'Cut compacted',
-                title: 'Cut selected contents, without indentation (Ctrl+Shift+X)',
-                onClick: () => handleCut(false),
-                disabled: !hasSelectionContents
-              }
-            ]
-          },
-          {
-            type: 'dropdown-button',
-            main: {
-              type: 'button',
-              onClick: () => handleCopy(true),
-              icon: faCopy,
-              text: 'Copy',
-              title: 'Copy selected contents, formatted with indentation (Ctrl+C)',
-              disabled: !hasSelectionContents
-            },
-            width: '12em',
-            items: [
-              {
-                type: 'button',
-                icon: faCopy,
-                text: 'Copy formatted',
-                title: 'Copy selected contents, formatted with indentation (Ctrl+C)',
-                onClick: () => handleCopy(true),
-                disabled: !hasSelectionContents
-              },
-              {
-                type: 'button',
-                icon: faCopy,
-                text: 'Copy compacted',
-                title: 'Copy selected contents, without indentation (Ctrl+Shift+C)',
-                onClick: () => handleCopy(false),
-                disabled: !hasSelectionContents
-              }
-            ]
-          },
-          {
-            type: 'button',
-            onClick: () => handlePasteFromMenu(),
-            icon: faPaste,
-            text: 'Paste',
-            title: 'Paste clipboard contents (Ctrl+V)',
-            disabled: !hasSelection
-          }
-        ]
-      },
-      { type: 'separator' },
-      {
-        type: 'row',
-        items: [
-          {
-            type: 'column',
-            items: [
-              {
-                type: 'button',
-                onClick: () => handleDuplicate(),
-                icon: faClone,
-                text: 'Duplicate',
-                title: 'Duplicate selected contents (Ctrl+D)',
-                disabled: !canDuplicate
-              },
-              {
-                type: 'button',
-                onClick: () => handleExtract(),
-                icon: faCropAlt,
-                text: 'Extract',
-                title: 'Extract selected contents',
-                disabled: !canExtract
-              },
-              {
-                type: 'button',
-                onClick: () => handleSortSelection(),
-                icon: faSortAmountDownAlt,
-                text: 'Sort',
-                title: 'Sort array or object contents',
-                disabled: !hasSelectionContents
-              },
-              {
-                type: 'button',
-                onClick: () => handleTransformSelection(),
-                icon: faFilter,
-                text: 'Transform',
-                title: 'Transform array or object contents (filter, sort, project)',
-                disabled: !hasSelectionContents
-              },
-              {
-                type: 'button',
-                onClick: () => handleRemove(),
-                icon: faTrashCan,
-                text: 'Remove',
-                title: 'Remove selected contents (Delete)',
-                disabled: !hasSelectionContents
-              }
-            ]
-          },
-          {
-            type: 'column',
-            items: [
-              { type: 'label', text: insertOrConvertText },
-              {
-                type: 'button',
-                onClick: () => handleInsertOrConvert('structure'),
-                icon: convertMode ? faArrowRightArrowLeft : faPlus,
-                text: 'Structure',
-                title: insertOrConvertText + ' structure',
-                disabled: !canInsertOrConvertStructure
-              },
-              {
-                type: 'button',
-                onClick: () => handleInsertOrConvert('object'),
-                icon: convertMode ? faArrowRightArrowLeft : faPlus,
-                text: 'Object',
-                title: insertOrConvertText + ' structure',
-                disabled: !canInsertOrConvertObject
-              },
-              {
-                type: 'button',
-                onClick: () => handleInsertOrConvert('array'),
-                icon: convertMode ? faArrowRightArrowLeft : faPlus,
-                text: 'Array',
-                title: insertOrConvertText + ' array',
-                disabled: !canInsertOrConvertArray
-              },
-              {
-                type: 'button',
-                onClick: () => handleInsertOrConvert('value'),
-                icon: convertMode ? faArrowRightArrowLeft : faPlus,
-                text: 'Value',
-                title: insertOrConvertText + ' value',
-                disabled: !canInsertOrConvertValue
-              }
-            ]
-          }
-        ]
-      },
-      {
-        type: 'separator'
-      },
-      {
-        type: 'row',
-        items: [
-          {
-            type: 'button',
-            onClick: () => handleInsertBefore(),
-            icon: faCaretSquareUp,
-            text: 'Insert before',
-            title: 'Select area before current entry to insert or paste contents',
-            disabled: !hasSelectionContents || rootSelected
-          },
-          {
-            type: 'button',
-            onClick: () => handleInsertAfter(),
-            icon: faCaretSquareDown,
-            text: 'Insert after',
-            title: 'Select area after current entry to insert or paste contents',
-            disabled: !hasSelectionContents || rootSelected
-          }
-        ]
-      }
-    ]
+    const defaultItems: ContextMenuItem[] = createTreeContextMenuItems({
+      json,
+      documentState,
+      parser,
+
+      onEditKey: handleEditKey,
+      onEditValue: handleEditValue,
+      onToggleEnforceString: handleToggleEnforceString,
+
+      onCut: handleCut,
+      onCopy: handleCopy,
+      onPaste: handlePasteFromMenu,
+
+      onRemove: handleRemove,
+      onDuplicate: handleDuplicate,
+      onExtract: handleExtract,
+
+      onInsertBefore: handleInsertBefore,
+      onInsert: handleInsertFromContextMenu,
+      onInsertAfter: handleInsertAfter,
+      onConvert: handleConvert,
+
+      onSort: handleSortSelection,
+      onTransform: handleTransformSelection
+    })
 
     const items = onRenderContextMenu(defaultItems)
 
     if (items === false) return
 
     const props = {
-      showTip,
+      tip: showTip
+        ? 'Tip: you can open this context menu via right-click or with Ctrl+Q'
+        : undefined,
       items,
       onCloseContextMenu: function () {
         closeAbsolutePopup(popupId)
@@ -2135,7 +1850,7 @@
 
     modalOpen = true
 
-    const popupId = openAbsolutePopup(TreeContextMenu, props, {
+    const popupId = openAbsolutePopup(ContextMenu, props, {
       left,
       top,
       offsetTop,
